Restructure codebase into feature-based directory layout
Moves all components and hooks from flat src/components/ and src/hooks/
into domain-driven src/features/<domain>/ directories:

  auth/        AuthScreen, EnableNotificationsScreen, ProfileSetupScreen,
               useAuth, usePushNotifications
  calendar/    CalendarView
  checks/      CheckActionsSheet, EditCheckModal, FirstCheckScreen,
               useChecks, useCheckComments
  events/      EditEventModal, EventCard, EventLobby, PasteModal, useEvents
  feed/        FeedView
  friends/     FriendsModal, UserProfileOverlay, useFriends
  notifications/ NotificationsPanel, useNotifications
  profile/     ProfileView
  squads/      GroupsView, SquadNotificationBanner, useSquads

App-shell pieces move to src/app/components/ and src/app/hooks/:
  components/  BottomNav, Header, Toast, DevProdBanner, UpdateBanner,
               IOSInstallBanner, Grain
  hooks/       useAppNavigation, usePullToRefresh, useToast

Generic utilities move to src/shared/hooks/:
  hooks/       useModalTransition

Deletes old css-module Button and LinkButton (superseded by
src/shared/components/Button using Tailwind).

Updates all import paths in page.tsx, layout.tsx, and admin/page.tsx.
