+++ src/features/auth/components/ProfileSetupScreen.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import * as db from "@/lib/db";
 import type { Profile } from "@/lib/types";
 import { font, color } from "@/lib/styles";
-import Grain from "./Grain";
+import Grain from "@/app/components/Grain";
 import { logError } from "@/lib/logger";
 
 const ProfileSetupScreen = ({
