+++ src/features/events/components/EventCard.tsx
