+++ src/features/squads/components/GroupsView.tsx
