+++ src/features/squads/components/SquadNotificationBanner.tsx
